ajax is working juhu!

+++ js/column.js
@@ -1,29 +1,55 @@
-function Column(name) {
+function Column(id, name) {
 	var self = this;
-
-	this.id = randomString();
-	this.name = name;
+	this.id = id;
+	this.name = name || 'No name given';
 	this.element = createColumn();
 
 	function createColumn() {
-		// TWORZENIE NOWYCH WĘZŁÓW
 		var column = $('<div class="column"></div>');
 		var columnTitle = $('<h2 class="column-title">' + self.name + '</h2>');
 		var columnCardList = $('<ul class="column-card-list"></ul>');
 		var columnDelete = $('<button class="btn-delete"><span class="fa fa-trash-o" aria-hidden="true"></span></button>');
 		var columnAddCard = $('<button class="column-add-card"><span class="fa fa-plus plus" aria-hidden="true"></span></button>');
 
-		// PODPINANIE ODPOWIEDNICH ZDARZEŃ POD WĘZŁY
+
 		columnDelete.click(function () {
-			self.deleteColumn();
+			alertify.confirm('Delete column', 'Hang out "' + self.name + '" column?', function () {
+				self.deleteColumn();
+				alertify.message('Column "' + name + '" deleted')
+			}, function () {
+				alertify.error('Uff... it was close!')
+			}).set('labels', {
+				ok: 'Hang up!',
+				cancel: 'Nope!'
+			});
 		});
 
-		columnAddCard.click(function () {
+		columnAddCard.click(function (event) {
+			alertify.prompt('Card name', 'Enter card name', 'Plese enter name here :)', function (evt, value) {
+				cardName = value;
+				$.ajax({
+					url: baseUrl + '/card',
+					method: 'POST',
+					data: {
+						name: cardName,
+						bootcamp_kanban_column_id: self.id
+					},
+					success: function (response) {
+						var card = new Card(response.id, cardName);
+						self.createCard(card);
+					}
+				});
+
+				alertify.success('Card "' + value + '" created');
+			}, function () {
+				alertify.error('Cancelled');
+			}).set('labels', {
+				ok: 'Done!',
+				cancel: 'I changed my mind'
+			});
 			event.preventDefault();
-			self.createCard(new Card(prompt("Wpisz nazwę karty")));
-		});  
-        
-		// KONSTRUOWANIE ELEMENTU KOLUMNY
+		});
+
 		column.append(columnTitle)
 			.append(columnDelete)
 			.append(columnAddCard)
@@ -31,12 +57,19 @@ function Column(name) {
 		return column;
 	}
 }
+
 Column.prototype = {
 	createCard: function (card) {
 		this.element.children('ul').append(card.element);
 	},
 	deleteColumn: function () {
-		this.element.remove();
+		var self = this;
+		$.ajax({
+			url: baseUrl + '/column/' + self.id,
+			method: 'DELETE',
+			success: function (response) {
+				self.element.remove();
+			}
+		});
 	}
 };
-
